SIEM/unparsed_logs_checker/requirements.txt
google-auth
python-dotenv
requests